Rename config, automatically merge pins
