+++ lib/config.js
@@ -83,6 +83,14 @@ const commitMessage = {
   commitMessageExtra: `from {{{currentVersion}}} ${getCommitMessageExtraDefault()}`
 };
 
+const autoMergePins = {
+  pin: {
+    masterIssueApproval: false,
+    enabled: true,
+    automerge: true
+  }
+};
+
 // -- END INDIVIDUAL CONFIG OPTIONS --------
 
 const shared = merge.all([
@@ -98,7 +106,8 @@ const shared = merge.all([
   issueApproval,
   ignoreEngineUpdates,
   orbUpdates,
-  artsyPRs
+  artsyPRs,
+  autoMergePins
 ]);
 
 const app = merge.all([
