+++ lib/generate-config.js
@@ -1,6 +1,6 @@
 const fs = require("fs");
 let pkg = require("../package.json");
-const { configs } = require("./config-builder");
+const { configs } = require("./config");
 
 pkg["renovate-config"] = configs;
